pyproject.toml
[tool.poetry]
name = "hyobot"
version = "0.1.0"
description = ""
authors = ["1kko-ahnlabio <1kko@ahnlab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.2"
python-dotenv = "^1.0.1"
fastapi = {extras = ["full"], version = "^0.111.0"}
uvicorn = {extras = ["standard"], version = "^0.30.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
